feat(engine/predicate): compiled predicates with cache; refactor gate to use it
